+++ CHANGES.md
@@ -65,6 +65,7 @@
 ## I/Os
 
 * Support for X source added (Java/Python) ([#X](https://github.com/apache/beam/issues/X)).
+* Support for Bigtable materialized view source added (Java) ([#38053](https://github.com/apache/beam/issues/38053)).
 
 ## New Features / Improvements